move out functions to separate files so its easier to import CloudPhind from clump_tools
